Forces draggable on children of ul that's draggable.
To do this, we need to give the `ul` a class name that we can select.
Previously we didn't modify the classnames of the `ul`. D3 will
subselect any element that is a child of the selection and matches the
selector, so the `selectAll('ul.draggable-list > li')` surprisingly
works.

Fixes #9